Create addActivities
Activities are the most important feature as they tell teachers HOW to teach specific points. 

They are all the most complicated so the video tutorials are essential. 

The main goal is to get teachers to use the "steps" feature to explain blow by blow how to run the activity. This should be highlighted in the main activity.