Create addLesson.eml
Lesson plans are simply a set of activities that are rolled together to achieve a given learning outcome. 

They are ideal ways to package all the things a teacher needs to teach a specific topic to their students. 

The instructions for lesson plans are contained in the activities so this high level object is more of a general guide that teachers can use to explain the finer points (theory, important nuances etc) that other teachers need to be aware of in order to maximise the learning outcomes.